rename: Lotto 파일 변경에 따른 import 수정
