+++ src/models/MyLotto.js
@@ -1,5 +1,5 @@
 import { Random } from '@woowacourse/mission-utils';
-import Lotto from './Lotto';
+import Lotto from '../Lotto';
 import NUMBER from '../constants/Number';
 
 class MyLotto {
